+++ TranslateCS2.Mod/Containers/Items/ModSettingsGroupGenerate.cs
@@ -26,18 +26,27 @@ internal partial class ModSettings {
     }
 
     [Exclude]
-    [SettingsUIButton]
     [SettingsUIDeveloper]
     [SettingsUISection(Section, GenerateGroup)]
     [SettingsUIDisableByCondition(typeof(ModSettings), nameof(IsGenerateLocalizationJsonHiddenDisabled))]
     [SettingsUIHideByCondition(typeof(ModSettings), nameof(IsGenerateLocalizationJsonHiddenDisabled))]
+    [SettingsUIDirectoryPicker]
+    public string GenerateDirectory { get; set; }
+
+    [Exclude]
+    [SettingsUIDeveloper]
+    [SettingsUISection(Section, GenerateGroup)]
+    [SettingsUIDisableByCondition(typeof(ModSettings), nameof(IsGenerateLocalizationJsonHiddenDisabled))]
+    [SettingsUIHideByCondition(typeof(ModSettings), nameof(IsGenerateLocalizationJsonHiddenDisabled))]
+    [SettingsUIConfirmation]
+    [SettingsUIButton]
     public bool GenerateLocalizationJson {
         set {
             if (this.SettingsLocale is null) {
                 return;
             }
             try {
-                string path = Path.Combine(this.runtimeContainer.Paths.ModsDataPathSpecific, ModConstants.ModExportKeyValueJsonName);
+                string path = Path.Combine(this.GenerateDirectory, ModConstants.ModExportKeyValueJsonName);
                 JsonHelper.Write(this.SettingsLocale.ExportableEntries, path);
             } catch (Exception ex) {
                 this.runtimeContainer.ErrorMessages.DisplayErrorMessageFailedToGenerateJson();
@@ -47,6 +56,7 @@ internal partial class ModSettings {
             }
         }
     }
+
     public bool IsGenerateLocalizationJsonHiddenDisabled() {
         return this.SettingsLocale is null;
     }